+++ src/ui.h
@@ -328,7 +328,7 @@ public:
         return input.pressed_keys() != KEY::A && hot_item == id && active_item == id;
     }
 
-    bool listbox(size_t *selected, const std::vector<std::string> &items) {
+    bool listbox(int *selected, const std::vector<std::string> &items) {
         *selected = clamp<int>(*selected, 0, items.size());
         const char *label = items[*selected].c_str();
         ui_id id = id_stack.get_id((void*)&items, sizeof(&items));
@@ -342,7 +342,7 @@ public:
         Rectangle<int> rect(xy, wh);
         new_selectable_widget(id, rect);
         if(hot_item == id && input.pressed_keys() == (KEY::UP | KEY::SELECT)) {
-            if(*selected < items.size() - 1)
+            if(*selected < (int)(items.size() - 1))
                 *selected += 1;
             active_item = id;
         } else if(hot_item == id && input.pressed_keys() == (KEY::DOWN | KEY::SELECT)) {
